hw/drivers/led: add led_itf
- Add concept of an led_itf
- Change directory structure to have hw/drivers/led
- Create a seperate package for hw/drivers/led
