+++ hw/drivers/led/lp5523/syscfg.yml
